feat: re-organize Nav in Self to Self, Options, FrontShelf, & Cognition.
