+++ app/Self/component/NavFront/SvgDoortoShelf.jsx
@@ -20,7 +20,7 @@ export default class SvgDoortoShelf extends React.Component {
         maxWidth: '100%',
         maxHeight: '100%'
       }}>
-      <defs><style>{".cls-1-toShelf{fill:#6c656c;stroke:#aaa;stroke-miterlimit:10;stroke-width:1.1px;}"}</style></defs>
+      <defs><style>{".cls-1-toShelf{fill:none;stroke:#aaa;stroke-miterlimit:10;stroke-width:1.1px;}"}</style></defs>
       <g id="圖層_2" data-name="圖層 2">
         <g id="圖層_5" data-name="圖層 5">
           <polyline className="cls-1-toShelf" points="21.61 26.51 21.61 29.51 21.61 3.9 5.39 0.55 25.9 0.55 25.9 25.51 25.9 24.93"/>
